requirements.txt
fastapi
uvicorn
boto3
python-multipart
google-cloud-firestore
firebase-admin
python-dotenv
